fix command dispatcher specs
